Return a masked array from the create3D function
This was placing zeros in every cell even when there was no data for
that cell.  This now returns masked values in the cells where there
is no data. I assumed that 0 could be an actual data value for some
variables and that I couldn't strip out all 0s later on in the process.