Add drawdown to available headertypes
The fdn-drawdown file specifies the following header:

1    1   1.000000E+00   1.000000E+00         DRAWDOWN   165   175     1 (10(1X1PE13.5))

The headertype in this case is drawdown which results in an error reading the file.

Specified drawdown type is not available. Available types are:
  1 head
  2 ucn

So I have added this type to the datafile-header types.
The dtype should be the same as from header_type head.