feat(pyshp 2): add support for writing shapefiles with pyshp 2
With pyshp 2, the call signature is different. This commit adds code that allows flopy to work with earlier versions of pyshp and with pyshp 2.  The code has now been tested with both pyshp 1.12 and pyshp 2.0.1.  There is a small remaining issue.  In some cases, we are passing three dimensional points to recarray2shp, but the z values are not and have never been written as a POINTZ type shapefile.  The shapefile writing should probably be modified in the future to have full support for z attributes.