feat(vtk): Added new htop argument to vtk writer
htop can now be passed into the vtk writer. htop should be the top of the cell or the head, whichever is lower. Also refactored the shared_vertex option so that when the ibound filter is used, only the points for the active cells are included in the vtk point list. Prior to this change, all of the vertex points for the grid were included in the vtk file, which resulted in very large vtk files.