feat(CellBudgetFile): method to return file position
Add method .get_position() to CellBudgetFile to return position of the
start of the data for a record in binary budget files. Optionally,
return the position of the start of the header for a record in binary
budget files. This method can be used to extract a slice from a binary
budget file. Update t017 autotest to test .get_position() method.

Modify .list_unique_records() and .list_unique_package_names() methods
to optionally convert byte strings to standard strings (using
.decode()). Update t017 and t039 autotests to accommodate this
modification.