Relax HOBs mlay Sum Criteria
Due to binary representation of base-10 floats in Python, an error was sometimes raised when reading in an HOB with multilayer observations. The sum of the layer contributions must equal 1; however, the sum could be calculated as 0.9999999 and the error would still be raised. This relaxes the criteria for the sum of layer contributions from multi-layer HOBs to, 0.99999 < sum < 1.00001. It also changes the error message to include the observation name where the error occurred.