fix(mfdis): multiple issues in mfdis resulted in incorrect row column determination (#560)
The ModflowDis Package has some intersection-like methods that are probably not used very widely, such as the get_rc_from_node_coordinates method.  The method did not seem to be tested very well for variable spacings along a column (DELC).  This fix eliminates code in the dis package and instead uses the intersection capabilities in the new modelgrid object.

Close #557