fix(Mflist): allow None as a list entry (#765)
* fix(Mflist): allow None, 0, and -1 as a list entry

Mflist did not allow a stress_period_dictionary to have None, 0, or -1 as a valid entry for a stress period.  But MODFLOW and MT3D do allow this.  Mflist was refactored so that a stress period dictionary could be spd = {0: None, 1: -1, 1: 0} or other combinations on this.  This is necessary for the mt3d ssm package, but it also was necessary for loading a valid well file that might have ITMP = -1 for all stress periods.  This would be a strange case for MODFLOW, but ITMP values of -1 for the first stress period of a wel package is valid for MODFLOW.

* slight modification for nss==0 in first period

closes #743