bug(flopy_io): limit fixed point format string to fixed column widths
flopy_io.write_fixed_var could write a fixed point string (f) that
exceeded the specified fixed column width. Added test to revert to
general format (g) with precision equal to column width - 6 to preserve
as much precision as possible. This value allows for a string with a
sign (+/-) and +/-e307. Added test for issue 1164 to t015_test.py
autotest.

Closes #1164