feat(GridIntersect): add optional kwarg to .intersect() to keep all cells with identical vertices
Add "keep_stacked_cells" kwarg to GridIntersect.intersect() which is used to prevent removal of
cells from the intersection results which have identical vertices as a cell already included in
the results. This kwarg is only used when intersecting using shapely, and is needed when dealing
with a DISU grid which doesn't have layers. With DIS and DISV, the intersection results include
cellids which have the layer number (either LAY, ROW, COL, or LAY, CELL2D) so cells above or
below the intersected cell can be easily found just be changing the layer number. But DISU
cellids do not include the layer so it is not at all easy to find cells above or below the
intersected cell. Thus having this kwarg which keeps all the cells in the results is needed.