fix(test_plot.py):: test_pathline_plot_xc
do not specify files, instead read from modflow object.  Previous version of modpath6 did not read the files but instead pulled the data from the model. Now if you want to specify the file you must pass the full path as it no longer assumes file is in the Modflow objects model_ws