Copied Variables from MODFLOW USG DISU to modelgrid
I am working with a MODFLOW USG DISU grid.  Not all of the variables that were read in from the grid are transferred to modelgrid.
I added code to do so.  In addition, when I was running the pytests on the original python code that I downloaded from Github, I
discovered some problems with a couple of the test files when running on Windows. I made changes to those files so that they would
run correctly.