chore: support python 3.13 (#2341)
Python 3.13 was recently released, which works fine with most of flopy's functionality.

VTK and PyVista do not yet support Python 3.13 from PyPI, as specified with conditional expressions. Pymetis is not yet available for 3.13 from conda-forge. Thus testing has been removed for now. For reference, use this doc to try and install Python and packages for 3.13.

Nevertheless, local testing on linux works with this PR for PyPI using pip install -e .[test,optional] which installs everything except VTK and PyVista.