ci(vtk): fix tests for vtk-9.4.0 by parsing XML file and checking attrs (#2380)
Fix a recent CI failure that uses the latest VTK package (9.4.0). The tests are changed from simple counting number of lines to checking attributes in the XML file using built-in XML tools, which should be more robust.

Binary wheels for vtk are now available for Python 3.13, so the conditional installs are removed for vtk and pyvista.