test(codegen): move codegen tests to ci job (#2412)
Test compatibility of the MF6 code generation mechanism with various MF6 versions in a CI job, rather than a pytest script. I think we can just run it daily, since the develop branch is tested on commit. Before this there was a CI issue which may have been related to invoking pip in a subprocess, which this now avoids.