fix(resolve_exe): allow shutil.which() to find exe without suffix in Windows (#2408)
This PR aims to address a behavior change in shutil.which (in Windows) introduced in python/cpython#127035.

For more context, in test_run_model_exe_rel_path, the mf6 executable is copied to a relative path, ../bin/mf6 (without extension). Within the test, shutil.which is used to locate the executable path. However, with the new changes in shutil.which, by default, it would only search for the path with an extension in PATHEXT, excluding paths without extension.