refactor(resolve_exe): also fix tests (#2464)
This extends a bit further than https://github.com/modflowpy/flopy/pull/2457 to resolve https://github.com/modflowpy/flopy/issues/2455 to use the str and PathLib versions of exe paths in the method. It also improves the docstring and return typing info.

But it also fixes autotest/test_mbase.py which was broken on Linux for paths that have mixed case to the repo. Paths don't need to be converted to lower() for equality comparisons, as this breaks mixed-case paths. The tests are written to compare pathlib objects, which works for the host path, e.g. Windows paths.

This also catches a missing warning thrown in autotest/test_modflow.py.