refactor(createpackages): use jinja for mf6 module code generation (#2333)
Reimplement createpackages.py with jinja. The aim is to match the old code generation capability without relying on mfstructure.py. This is a first step toward a leaner python representation of the mf6 input specification.

Note: mfstructure.py is still used at runtime which will need to be unraveled in followup work.

The new code generation machinery lives in flopy.mf6.utils.codegen. It includes a bunch of filters handling quirks of the generated classes which we can aim to eliminate in future. The templates should also get much simpler in future.

The module now consumes TOML rather than legacy DFNs. modflow-devtools is used at code generation time to convert DFNs to TOML, then generate sources as before. The conversion happens behind the scenes and does not change anything for the user of the codegen utility.

A new optional dependency group 'codegen' is introduced, with Jinja2 and modflow-devtools, which are required for the code generation utilities. These and some transitive dependencies (tomli/tomli_w) are added to environment.yml.

Miscellaneous:

- expand and add some mermaid diagrams to the MF6 module dev guide
- minor fix in flopy/mf6/data/mfdatastorage.py to avoid referencing a variable before it exists
- update DFNs as per MODFLOW-ORG/modflow6#2031